requirements.txt
# Core Streamlit App
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0

# AI/ML Core
scikit-learn>=1.3.0
torch>=2.0.0

# Cellpose (GPU optional on cloud)
cellpose>=3.0.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Medical Data
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.2
